PR ready changes

+++ run_demo.py
@@ -15,12 +15,12 @@ import argparse
 if __name__=='__main__':
   parser = argparse.ArgumentParser()
   code_dir = os.path.dirname(os.path.realpath(__file__))
-  parser.add_argument('--mesh_file', type=str, default='/home/tushar/Desktop/FoundationPoseForSpotSim2Real/demo_data/bottle_scan_anchor_intel/mesh/scaled_with_pytorch3d/mesh.obj')
-  parser.add_argument('--test_scene_dir', type=str, default=f'{code_dir}/demo_data/bottlevideocloser')
+  parser.add_argument('--mesh_file', type=str, default=f'{code_dir}/demo_data/mustard0/mesh/textured_simple.obj')
+  parser.add_argument('--test_scene_dir', type=str, default=f'{code_dir}/demo_data/mustard0')
   parser.add_argument('--est_refine_iter', type=int, default=5)
   parser.add_argument('--track_refine_iter', type=int, default=2)
-  parser.add_argument('--debug', type=int, default=1)
-  parser.add_argument('--debug_dir', type=str, default=f'{code_dir}/demo_data/bottlevideocloser/debug')
+  parser.add_argument('--debug', type=int, default=3)
+  parser.add_argument('--debug_dir', type=str, default=f'{code_dir}/demo_data/mustard0/debug')
   args = parser.parse_args()
 
   set_logging_format()
@@ -30,7 +30,11 @@ if __name__=='__main__':
 
   debug = args.debug
   debug_dir = args.debug_dir
-  os.system(f'rm -rf {debug_dir}/* && mkdir -p {debug_dir}/track_vis {debug_dir}/ob_in_cam')
+  
+  if os.path.exists(debug_dir):
+    os.system(f'rm -rf {debug_dir}/* && mkdir -p {debug_dir}/track_vis {debug_dir}/ob_in_cam')
+  else:
+    os.system(f'mkdir -p {debug_dir}/track_vis {debug_dir}/ob_in_cam')
 
   to_origin, extents = trimesh.bounds.oriented_bounds(mesh)
   bbox = np.stack([-extents/2, extents/2], axis=0).reshape(2,3)
